+++ storybook/src/components/form/form-group-select.css-template.ts
@@ -67,7 +67,7 @@ export const cssFormGroupSelect: ComponentImplementation<FormGroupSelect<Templat
                 `
               : nothing}
             ${formGroup.errorText && formGroup.state === "invalid"
-              ? html`<p class="dso-message" id=${errorTextId}>${formGroup.errorText}</p>`
+              ? html`<p class="dso-message" role="alert" id=${errorTextId}>${formGroup.errorText}</p>`
               : nothing}
             ${formGroup.helpText ? html`<p class="dso-help-block" id=${helpTextId}>${formGroup.helpText}</p>` : nothing}
           </div>