+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+### Fixed
+* Form: Fix error messages ([#2141](https://github.com/dso-toolkit/dso-toolkit/issues/2141))
+
 ### Changed
 * Form Group: Errormessage error icon ([#2640](https://github.com/dso-toolkit/dso-toolkit/issues/2640))